Add request header filter support for gRPC
