+++ internal/mode/static/state/graph/httproute_test.go
@@ -1189,325 +1189,3 @@ func TestValidateFilterRewrite(t *testing.T) {
 		})
 	}
 }
-
-func TestValidateFilterRequestHeaderModifier(t *testing.T) {
-	createAllValidValidator := func() *validationfakes.FakeHTTPFieldsValidator {
-		v := &validationfakes.FakeHTTPFieldsValidator{}
-		return v
-	}
-
-	tests := []struct {
-		filter         gatewayv1.HTTPRouteFilter
-		validator      *validationfakes.FakeHTTPFieldsValidator
-		name           string
-		expectErrCount int
-	}{
-		{
-			validator: createAllValidValidator(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterRequestHeaderModifier,
-				RequestHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Set: []gatewayv1.HTTPHeader{
-						{Name: "MyBespokeHeader", Value: "my-value"},
-					},
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "Accept-Encoding", Value: "gzip"},
-					},
-					Remove: []string{"Cache-Control"},
-				},
-			},
-			expectErrCount: 0,
-			name:           "valid request header modifier filter",
-		},
-		{
-			validator: createAllValidValidator(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type:                  gatewayv1.HTTPRouteFilterRequestHeaderModifier,
-				RequestHeaderModifier: nil,
-			},
-			expectErrCount: 1,
-			name:           "nil request header modifier filter",
-		},
-		{
-			validator: func() *validationfakes.FakeHTTPFieldsValidator {
-				v := createAllValidValidator()
-				v.ValidateFilterHeaderNameReturns(errors.New("Invalid header"))
-				return v
-			}(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterRequestHeaderModifier,
-				RequestHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "$var_name", Value: "gzip"},
-					},
-				},
-			},
-			expectErrCount: 1,
-			name:           "request header modifier filter with invalid add",
-		},
-		{
-			validator: func() *validationfakes.FakeHTTPFieldsValidator {
-				v := createAllValidValidator()
-				v.ValidateFilterHeaderNameReturns(errors.New("Invalid header"))
-				return v
-			}(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterRequestHeaderModifier,
-				RequestHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Remove: []string{"$var-name"},
-				},
-			},
-			expectErrCount: 1,
-			name:           "request header modifier filter with invalid remove",
-		},
-		{
-			validator: func() *validationfakes.FakeHTTPFieldsValidator {
-				v := createAllValidValidator()
-				v.ValidateFilterHeaderValueReturns(errors.New("Invalid header value"))
-				return v
-			}(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterRequestHeaderModifier,
-				RequestHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "Accept-Encoding", Value: "yhu$"},
-					},
-				},
-			},
-			expectErrCount: 1,
-			name:           "request header modifier filter with invalid header value",
-		},
-		{
-			validator: func() *validationfakes.FakeHTTPFieldsValidator {
-				v := createAllValidValidator()
-				v.ValidateFilterHeaderValueReturns(errors.New("Invalid header value"))
-				v.ValidateFilterHeaderNameReturns(errors.New("Invalid header"))
-				return v
-			}(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterRequestHeaderModifier,
-				RequestHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Set: []gatewayv1.HTTPHeader{
-						{Name: "Host", Value: "my_host"},
-					},
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "}90yh&$", Value: "gzip$"},
-						{Name: "}67yh&$", Value: "compress$"},
-					},
-					Remove: []string{"Cache-Control$}"},
-				},
-			},
-			expectErrCount: 7,
-			name:           "request header modifier filter all fields invalid",
-		},
-		{
-			validator: createAllValidValidator(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterRequestHeaderModifier,
-				RequestHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Set: []gatewayv1.HTTPHeader{
-						{Name: "MyBespokeHeader", Value: "my-value"},
-						{Name: "mYbespokeHEader", Value: "duplicate"},
-					},
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "Accept-Encoding", Value: "gzip"},
-						{Name: "accept-encodING", Value: "gzip"},
-					},
-					Remove: []string{"Cache-Control", "cache-control"},
-				},
-			},
-			expectErrCount: 3,
-			name:           "request header modifier filter not unique names",
-		},
-	}
-
-	filterPath := field.NewPath("test")
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			g := NewWithT(t)
-			allErrs := validateFilterHeaderModifier(
-				test.validator, test.filter.RequestHeaderModifier, filterPath,
-			)
-			g.Expect(allErrs).To(HaveLen(test.expectErrCount))
-		})
-	}
-}
-
-func TestValidateFilterResponseHeaderModifier(t *testing.T) {
-	createAllValidValidator := func() *validationfakes.FakeHTTPFieldsValidator {
-		v := &validationfakes.FakeHTTPFieldsValidator{}
-		return v
-	}
-
-	tests := []struct {
-		filter         gatewayv1.HTTPRouteFilter
-		validator      *validationfakes.FakeHTTPFieldsValidator
-		name           string
-		expectErrCount int
-	}{
-		{
-			validator: createAllValidValidator(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Set: []gatewayv1.HTTPHeader{
-						{Name: "MyBespokeHeader", Value: "my-value"},
-					},
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "Accept-Encoding", Value: "gzip"},
-					},
-					Remove: []string{"Cache-Control"},
-				},
-			},
-			expectErrCount: 0,
-			name:           "valid response header modifier filter",
-		},
-		{
-			validator: createAllValidValidator(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type:                   gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: nil,
-			},
-			expectErrCount: 1,
-			name:           "nil response header modifier filter",
-		},
-		{
-			validator: func() *validationfakes.FakeHTTPFieldsValidator {
-				v := createAllValidValidator()
-				v.ValidateFilterHeaderNameReturns(errors.New("Invalid header"))
-				return v
-			}(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "$var_name", Value: "gzip"},
-					},
-				},
-			},
-			expectErrCount: 1,
-			name:           "response header modifier filter with invalid add",
-		},
-		{
-			validator: func() *validationfakes.FakeHTTPFieldsValidator {
-				v := createAllValidValidator()
-				v.ValidateFilterHeaderNameReturns(errors.New("Invalid header"))
-				return v
-			}(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Remove: []string{"$var-name"},
-				},
-			},
-			expectErrCount: 1,
-			name:           "response header modifier filter with invalid remove",
-		},
-		{
-			validator: func() *validationfakes.FakeHTTPFieldsValidator {
-				v := createAllValidValidator()
-				v.ValidateFilterHeaderValueReturns(errors.New("Invalid header value"))
-				return v
-			}(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "Accept-Encoding", Value: "yhu$"},
-					},
-				},
-			},
-			expectErrCount: 1,
-			name:           "response header modifier filter with invalid header value",
-		},
-		{
-			validator: func() *validationfakes.FakeHTTPFieldsValidator {
-				v := createAllValidValidator()
-				v.ValidateFilterHeaderValueReturns(errors.New("Invalid header value"))
-				v.ValidateFilterHeaderNameReturns(errors.New("Invalid header"))
-				return v
-			}(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Set: []gatewayv1.HTTPHeader{
-						{Name: "Host", Value: "my_host"},
-					},
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "}90yh&$", Value: "gzip$"},
-						{Name: "}67yh&$", Value: "compress$"},
-					},
-					Remove: []string{"Cache-Control$}"},
-				},
-			},
-			expectErrCount: 7,
-			name:           "response header modifier filter all fields invalid",
-		},
-		{
-			validator: createAllValidValidator(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Set: []gatewayv1.HTTPHeader{
-						{Name: "MyBespokeHeader", Value: "my-value"},
-						{Name: "mYbespokeHEader", Value: "duplicate"},
-					},
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "Accept-Encoding", Value: "gzip"},
-						{Name: "accept-encodING", Value: "gzip"},
-					},
-					Remove: []string{"Cache-Control", "cache-control"},
-				},
-			},
-			expectErrCount: 3,
-			name:           "response header modifier filter not unique names",
-		},
-		{
-			validator: createAllValidValidator(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Set: []gatewayv1.HTTPHeader{
-						{Name: "Content-Length", Value: "163"},
-					},
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "Content-Type", Value: "text/plain"},
-					},
-					Remove: []string{"X-Pad"},
-				},
-			},
-			expectErrCount: 3,
-			name:           "response header modifier filter with disallowed header name",
-		},
-		{
-			validator: createAllValidValidator(),
-			filter: gatewayv1.HTTPRouteFilter{
-				Type: gatewayv1.HTTPRouteFilterResponseHeaderModifier,
-				ResponseHeaderModifier: &gatewayv1.HTTPHeaderFilter{
-					Set: []gatewayv1.HTTPHeader{
-						{Name: "X-Accel-Redirect", Value: "/protected/iso.img"},
-					},
-					Add: []gatewayv1.HTTPHeader{
-						{Name: "X-Accel-Limit-Rate", Value: "1024"},
-					},
-					Remove: []string{"X-Accel-Charset"},
-				},
-			},
-			expectErrCount: 3,
-			name:           "response header modifier filter with disallowed header name prefix",
-		},
-	}
-
-	filterPath := field.NewPath("test")
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			g := NewWithT(t)
-			allErrs := validateFilterResponseHeaderModifier(
-				test.validator, test.filter.ResponseHeaderModifier, filterPath,
-			)
-			g.Expect(allErrs).To(HaveLen(test.expectErrCount))
-		})
-	}
-}
