+++ internal/mode/static/state/graph/route_common.go
@@ -573,3 +573,162 @@ func validateHeaderMatch(
 
 	return allErrs
 }
+
+func validateFilterHeaderModifier(
+	validator validation.HTTPFieldsValidator,
+	headerModifier *v1.HTTPHeaderFilter,
+	filterPath *field.Path,
+) field.ErrorList {
+	if headerModifier == nil {
+		return field.ErrorList{field.Required(filterPath, "cannot be nil")}
+	}
+
+	return validateFilterHeaderModifierFields(validator, headerModifier, filterPath)
+}
+
+func validateFilterHeaderModifierFields(
+	validator validation.HTTPFieldsValidator,
+	headerModifier *v1.HTTPHeaderFilter,
+	headerModifierPath *field.Path,
+) field.ErrorList {
+	var allErrs field.ErrorList
+
+	// Ensure that the header names are case-insensitive unique
+	allErrs = append(allErrs, validateRequestHeadersCaseInsensitiveUnique(
+		headerModifier.Add,
+		headerModifierPath.Child(add))...,
+	)
+	allErrs = append(allErrs, validateRequestHeadersCaseInsensitiveUnique(
+		headerModifier.Set,
+		headerModifierPath.Child(set))...,
+	)
+	allErrs = append(allErrs, validateRequestHeaderStringCaseInsensitiveUnique(
+		headerModifier.Remove,
+		headerModifierPath.Child(remove))...,
+	)
+
+	for _, h := range headerModifier.Add {
+		if err := validator.ValidateFilterHeaderName(string(h.Name)); err != nil {
+			valErr := field.Invalid(headerModifierPath.Child(add), h, err.Error())
+			allErrs = append(allErrs, valErr)
+		}
+		if err := validator.ValidateFilterHeaderValue(h.Value); err != nil {
+			valErr := field.Invalid(headerModifierPath.Child(add), h, err.Error())
+			allErrs = append(allErrs, valErr)
+		}
+	}
+	for _, h := range headerModifier.Set {
+		if err := validator.ValidateFilterHeaderName(string(h.Name)); err != nil {
+			valErr := field.Invalid(headerModifierPath.Child(set), h, err.Error())
+			allErrs = append(allErrs, valErr)
+		}
+		if err := validator.ValidateFilterHeaderValue(h.Value); err != nil {
+			valErr := field.Invalid(headerModifierPath.Child(set), h, err.Error())
+			allErrs = append(allErrs, valErr)
+		}
+	}
+	for _, h := range headerModifier.Remove {
+		if err := validator.ValidateFilterHeaderName(h); err != nil {
+			valErr := field.Invalid(headerModifierPath.Child(remove), h, err.Error())
+			allErrs = append(allErrs, valErr)
+		}
+	}
+
+	return allErrs
+}
+
+func validateFilterResponseHeaderModifier(
+	validator validation.HTTPFieldsValidator,
+	responseHeaderModifier *v1.HTTPHeaderFilter,
+	filterPath *field.Path,
+) field.ErrorList {
+	if errList := validateFilterHeaderModifier(validator, responseHeaderModifier, filterPath); errList != nil {
+		return errList
+	}
+	var allErrs field.ErrorList
+
+	allErrs = append(allErrs, validateResponseHeaders(
+		responseHeaderModifier.Add,
+		filterPath.Child(add))...,
+	)
+
+	allErrs = append(allErrs, validateResponseHeaders(
+		responseHeaderModifier.Set,
+		filterPath.Child(set))...,
+	)
+
+	var removeHeaders []v1.HTTPHeader
+	for _, h := range responseHeaderModifier.Remove {
+		removeHeaders = append(removeHeaders, v1.HTTPHeader{Name: v1.HTTPHeaderName(h)})
+	}
+
+	allErrs = append(allErrs, validateResponseHeaders(
+		removeHeaders,
+		filterPath.Child(remove))...,
+	)
+
+	return allErrs
+}
+
+func validateResponseHeaders(
+	headers []v1.HTTPHeader,
+	path *field.Path,
+) field.ErrorList {
+	var allErrs field.ErrorList
+	disallowedResponseHeaderSet := map[string]struct{}{
+		"server":         {},
+		"date":           {},
+		"x-pad":          {},
+		"content-type":   {},
+		"content-length": {},
+		"connection":     {},
+	}
+	invalidPrefix := "x-accel"
+
+	for _, h := range headers {
+		valErr := field.Invalid(path, h, "header name is not allowed")
+		name := strings.ToLower(string(h.Name))
+		if _, exists := disallowedResponseHeaderSet[name]; exists || strings.HasPrefix(name, strings.ToLower(invalidPrefix)) {
+			allErrs = append(allErrs, valErr)
+		}
+	}
+
+	return allErrs
+}
+
+func validateRequestHeadersCaseInsensitiveUnique(
+	headers []v1.HTTPHeader,
+	path *field.Path,
+) field.ErrorList {
+	var allErrs field.ErrorList
+
+	seen := make(map[string]struct{})
+
+	for _, h := range headers {
+		name := strings.ToLower(string(h.Name))
+		if _, exists := seen[name]; exists {
+			valErr := field.Invalid(path, h, "header name is not unique")
+			allErrs = append(allErrs, valErr)
+		}
+		seen[name] = struct{}{}
+	}
+
+	return allErrs
+}
+
+func validateRequestHeaderStringCaseInsensitiveUnique(headers []string, path *field.Path) field.ErrorList {
+	var allErrs field.ErrorList
+
+	seen := make(map[string]struct{})
+
+	for _, h := range headers {
+		name := strings.ToLower(h)
+		if _, exists := seen[name]; exists {
+			valErr := field.Invalid(path, h, "header name is not unique")
+			allErrs = append(allErrs, valErr)
+		}
+		seen[name] = struct{}{}
+	}
+
+	return allErrs
+}
