+++ site/content/overview/gateway-api-compatibility.md
@@ -199,7 +199,10 @@ See the [static-mode]({{< relref "/reference/cli-help.md#static-mode">}}) comman
     - `matches`
       - `method`: Partially supported. Only `Exact` type with both `method.service` and `method.method` specified.
       - `headers`: Partially supported. Only `Exact` type.
-    - `filters`: Not supported
+    - `filters`
+      - `type`: Supported.
+      - `requestHeaderModifier`: Supported. If multiple filters are configured, NGINX Gateway Fabric will choose the first and ignore the rest.
+      - `responseHeaderModifier`, `requestMirror`, `extensionRef`: Not supported.
     - `backendRefs`: Partially supported. Backend ref `filters` are not supported.
 - `status`
   - `parents`
